# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Yamanouchi Snow Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /weather:
    get:
      operationId: getWeather
      tags: [weather]
      summary: Get weather for all resort locations
      description: Returns current weather and 7-day forecast for Thredbo, Perisher, Charlotte's Pass, and Jindabyne
      responses:
        "200":
          description: Weather data for all locations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeatherResponse"
        "500":
          description: Error fetching weather
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /weather/{locationId}:
    get:
      operationId: getLocationWeather
      tags: [weather]
      summary: Get weather for a specific location
      description: Returns current weather and 7-day forecast for a specific resort location
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: string
            enum: [thredbo, perisher, charlottes-pass, jindabyne]
      responses:
        "200":
          description: Weather data for the location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationWeather"
        "404":
          description: Location not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error fetching weather
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    CurrentWeather:
      type: object
      properties:
        temperature:
          type: number
        feelsLike:
          type: number
        humidity:
          type: number
        windSpeed:
          type: number
        windDirection:
          type: number
        windDirectionCompass:
          type: string
        windGust:
          type: number
        weatherCode:
          type: integer
        weatherDescription:
          type: string
        isDay:
          type: boolean
        snowDepth:
          type: number
        precipitation:
          type: number
        cloudCover:
          type: integer
        visibility:
          type: number
        pressure:
          type: number
        dewpoint:
          type: number
        rainSince9am:
          type: number
        dataSource:
          type: string
        bomStation:
          type: string
        bomObservationTime:
          type: string
      required:
        - temperature
        - feelsLike
        - humidity
        - windSpeed
        - windDirection
        - weatherCode
        - weatherDescription
        - isDay
        - precipitation
        - cloudCover
    DailyForecast:
      type: object
      properties:
        date:
          type: string
        maxTemp:
          type: number
        minTemp:
          type: number
        weatherCode:
          type: integer
        weatherDescription:
          type: string
        precipitationSum:
          type: number
        snowfallSum:
          type: number
        windSpeedMax:
          type: number
        uvIndexMax:
          type: number
        sunrise:
          type: string
        sunset:
          type: string
      required:
        - date
        - maxTemp
        - minTemp
        - weatherCode
        - weatherDescription
        - precipitationSum
        - snowfallSum
        - windSpeedMax
    HourlyForecast:
      type: object
      properties:
        time:
          type: string
        temperature:
          type: number
        weatherCode:
          type: integer
        weatherDescription:
          type: string
        precipitation:
          type: number
        snowfall:
          type: number
        windSpeed:
          type: number
        humidity:
          type: number
        feelsLike:
          type: number
        cloudCover:
          type: integer
      required:
        - time
        - temperature
        - weatherCode
        - weatherDescription
        - precipitation
        - windSpeed
        - humidity
        - feelsLike
    LocationInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        elevation:
          type: number
        latitude:
          type: number
        longitude:
          type: number
        description:
          type: string
        bomStation:
          type: string
        bomStationId:
          type: string
      required:
        - id
        - name
        - elevation
        - latitude
        - longitude
        - description
    LocationWeather:
      type: object
      properties:
        location:
          $ref: "#/components/schemas/LocationInfo"
        current:
          $ref: "#/components/schemas/CurrentWeather"
        daily:
          type: array
          items:
            $ref: "#/components/schemas/DailyForecast"
        hourly:
          type: array
          items:
            $ref: "#/components/schemas/HourlyForecast"
        lastUpdated:
          type: string
      required:
        - location
        - current
        - daily
        - hourly
        - lastUpdated
    WeatherResponse:
      type: object
      properties:
        locations:
          type: array
          items:
            $ref: "#/components/schemas/LocationWeather"
        lastUpdated:
          type: string
      required:
        - locations
        - lastUpdated
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message