# lib/api-zod/src/generated/types/currentWeather.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Yamanouchi Snow Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CurrentWeather {
  temperature: number;
  feelsLike: number;
  humidity: number;
  windSpeed: number;
  windDirection: number;
  windDirectionCompass?: string;
  windGust?: number;
  weatherCode: number;
  weatherDescription: string;
  isDay: boolean;
  snowDepth?: number;
  precipitation: number;
  cloudCover: number;
  visibility?: number;
  pressure?: number;
  dewpoint?: number;
  rainSince9am?: number;
  dataSource?: string;
  bomStation?: string;
  bomObservationTime?: string;
}