# lib/api-zod/src/generated/types/locationInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Yamanouchi Snow Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface LocationInfo {
  id: string;
  name: string;
  elevation: number;
  latitude: number;
  longitude: number;
  description: string;
  bomStation?: string;
  bomStationId?: string;
}